initial commite.
Migrating from the subversion history was a PITA so this is just
a fresh import. If you really really need the history then a dump
of the svn repo is in the backups. You can restore that to a new
svn repo and review the history there. Although, I know the history
is not too interesting for this repo.